refactor: changed app route group to countdown route group
