+++ src/modules/countdown/ui/components/countdown-nav-dropdown.tsx
@@ -13,7 +13,7 @@ import { ChevronDown, LayoutDashboard, Plus } from "lucide-react";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
 
-export const AppNavDropdown = () => {
+export const CountdownNavDropdown = () => {
   const pathname = usePathname();
 
   // Temporary mock data until connected to DB
@@ -23,26 +23,23 @@ export const AppNavDropdown = () => {
     { id: "3", name: "Winter Break" },
   ];
 
-  // Get current page name based on pathname
   const getCurrentPageName = () => {
     if (pathname === "/dashboard")
       return (
         <>
-          {/* <LayoutDashboard className="h-4 w-4" /> */}
           <span className="truncate">Dashboard</span>
         </>
       );
     if (pathname === "/countdown/new")
       return (
         <>
-          {/* <Plus className="h-4 w-4" /> */}
           <span className="truncate">New Countdown</span>
         </>
       );
 
     const countdownId = pathname.split("/").pop();
     const currentCountdown = countdowns.find((c) => c.id === countdownId);
-    // Wrap the countdown name or default text in a span with truncate
+
     return (
       <span className="truncate">
         {currentCountdown?.name ?? "Select Countdown"}
