+++ src/modules/countdown/ui/components/countdown-nav-items.tsx
@@ -1,9 +1,8 @@
 import { ClerkUserButton } from "@/modules/auth/ui/components/clerk-user-button";
 import { auth } from "@clerk/nextjs/server";
 import { redirect } from "next/navigation";
-import { AppNavDropdown } from "./app-nav-dropdown";
 
-export const AppNavItems = async () => {
+export const CountdownNavItems = async () => {
   const { userId } = await auth();
 
   if (!userId) {
@@ -12,9 +11,7 @@ export const AppNavItems = async () => {
 
   return (
     <div className="flex items-center gap-2">
-      <div className="hidden sm:flex">
-        <AppNavDropdown />
-      </div>
+      <div className="hidden sm:flex">{/* <CountdownNavDropdown /> */}</div>
       <div className="h-8 w-8">
         <ClerkUserButton />
       </div>
